refactor(env): P4 - add env-shared public facade and migrate consumer imports
- Create env-shared.ts as a public re-export facade over env-files.ts
- Migrate 11 consumer files from env-files.js to env-shared.js imports
